ADD: Papelera funcional con BUGS y implementacion en backend de eliminado o mover a papelera
Falta solucionar un bug de error en los paths al momento de cambiar en el sidebar y implementar el botón de eliminado en las carpetas

+++ backend/cloud/utils/sys_management.py
@@ -90,6 +90,15 @@ class SysManagement:
         tree_structure = StructureCurrentPath(current_path=base_path)
         tree_structure.current_items_on_path()
         return jsonable_encoder(tree_structure.current_items)
+    
+    async def get_current_level_trash(self , path_on_folder : Optional[Path] = None) -> Dict[str , Any]:
+        if path_on_folder is None:
+            base_path = self.cloud_builder.current_path / '.trash/'
+        else:
+            base_path = self.cloud_builder.current_path / '.trash/' / path_on_folder
+        tree_structure = StructureCurrentPath(current_path=base_path)
+        tree_structure.current_items_on_path()
+        return jsonable_encoder(tree_structure.current_items)
 
     async def create_dir(self , name_new_dir : str , path_on_folder : Optional[str] = None):
         """
@@ -157,24 +166,28 @@ class SysManagement:
             raise HTTPException(status_code=404, detail="El archivo ya ha sido eliminado o no existe")
 
         
-    def delete_folder(self , path_folder : Optional[Path] = None) -> float:
+    async def delete_or_mv_folder(self , path_folder : Optional[Path] = None) -> Dict[str , bool]:
         if path_folder is None:
             raise Exception("Set a folder")
         base_path = self.cloud_builder.build_path(path_folder)
         if not base_path.is_dir():
             raise Exception("This path pointer not folder")
         try:
+            #----------- Logica de eliminado de carpetas vacias del sistema. ------------------
             structure = StructureCurrentPath(current_path=base_path)
             structure.current_items_on_path()
-            shutil.rmtree(str(base_path))
-            free_space = sum(item.size for item in structure.current_items)
-            return jsonable_encoder({"tree_rm" : structure.current_items, "free" : free_space })
+            if len(structure.current_items) != 0:
+                await self.__run_async_command(f"mv {base_path} {self.cloud_builder.current_path}/.trash/")
+                return {"deleted" : False , "moved" : True}
+            else:
+                shutil.rmtree(str(base_path))
+                return {"deleted" : True , "moved" : False}
+            #----------------------------------------------------------------------------------
         except FileNotFoundError:
             raise HTTPException(status_code=404, detail="El directorio ya ha sido eliminado o no existe")
 
         except PermissionError:
             raise HTTPException(status_code=403 , detail="No posees permisos")
-        ...
     
     def __generate_uuid_symlink(self) -> uuid.UUID:
         return uuid.uuid4()
@@ -251,8 +264,8 @@ class SysManagement:
             zip_name = resource_path.name + ".zip"
             zip_path_destino = symlink_path_destination / Path(f"{zip_name}") #Para que se cree dentro de la carpeta de transferidos del usuario receptor
             with zipfile.ZipFile(zip_path_destino , "w") as zipf:
-                for file in resource_path.glob("*"):
-                    zipf.write(file , arcname=file.name)
+                for file in resource_path.rglob("*"):
+                    zipf.write(file , arcname=file.relative_to(resource_path))
         else:
             symlink_path_destination_file = symlink_path_destination / f"{str(resource_path.name)}"
             shutil.copyfile(resource_path , symlink_path_destination_file)
